backend/requirements.txt
Flask
flask-cors
faiss-cpu
numpy
sentence-transformers
torch
PyPDF2
